chore(api): divide API code into different files
